add the different disease informations for the differents profiles
