+++ app/pharmaXcess_app/screens/health/DiseasesProfile.native.tsx
@@ -3,35 +3,191 @@ import { View, Text, TouchableOpacity, ScrollView, Alert, Modal, TextInput } fro
 import { LinearGradient } from 'expo-linear-gradient';
 import { StackNavigationProp } from '@react-navigation/stack';
 import { Ionicons } from '@expo/vector-icons';
-import createStyles from '../../styles/CardGrid.style';
+import createStyles from '../../styles/ProfileInfos.style';
 import { useTheme } from '../../context/ThemeContext';
 import { useFontScale } from '../../context/FontScaleContext';
 import { useProfile } from '../../context/ProfileContext';
 import { useProfileData } from '../../hooks/useProfileData';
+import { CustomPicker } from '../../components';
+
+type Disease = { 
+    name: string; 
+    description: string; 
+    symptoms: string; 
+    beginDate: string; 
+    medications: string; 
+    examens: string; 
+}; 
 
 type DiseasesProps = { 
     navigation: StackNavigationProp<any, any>; 
 }; 
 
-export default function Diseases({ navigation }: DiseasesProps): React.JSX.Element {
+export default function Diseases ({ navigation }: DiseasesProps): React.JSX.Element {
     const { colors } = useTheme();
     const { fontScale } = useFontScale();
     const { currentProfile } = useProfile();
-    const { diseases, addDisease, removeDisease } = useProfileData();
+    const { diseases: profileDiseases, addDisease, removeDisease } = useProfileData();
     const styles = createStyles(colors, fontScale);
 
-    const [isModalVisible, setModalVisible] = useState<boolean>(false);
-    const [newDisease, setNewDisease] = useState<string>('');
+    // Diseases predefined for the main profile 
+    const [diseases, setDiseases] = useState<Disease[]>([
+        {
+            name: 'Diabète',
+            description: 'Le diabète est une maladie chronique qui se caractérise par un excès de sucre dans le sang.',
+            symptoms: 'soif intense, besoin fréquent d\'uriner, fatigue, perte de poids, vision floue, cicatrisation lente, infections fréquentes, démangeaisons, fourmillements, douleurs, crampes, nausées, vomissements, haleine fruitée, perte de conscience',
+            beginDate: '01/01/2000',
+            medications: 'insuline, metformine, sulfamide hypoglycémiants, glinides, glitazones, inhibiteurs de l\'alpha-glucosidase, inhibiteurs de la DPP-4, agonistes des récepteurs du GLP-1, inhibiteurs du cotransporteur du sodium-glucose de type 2',
+            examens: 'glycémie à jeun, hémoglobine glyquée, test de tolérance au glucose, test de glycémie aléatoire, test de glycémie postprandiale (après un repas)',
+        },
+        {
+            name: 'Hypertension',
+            description: 'L\'hypertension artérielle est une maladie chronique caractérisée par une pression artérielle trop élevée dans les artères.',
+            symptoms: 'maux de tête, fatigue, étourdissements, bourdonnements d\'oreilles, palpitations, douleurs thoraciques, essoufflement, saignements de nez, vision floue',
+            beginDate: '01/01/2005',
+            medications: 'diurétiques, bêta-bloquants, inhibiteurs de l\'enzyme de conversion de l\'angiotensine (IECA), antagonistes des récepteurs de l\'angiotensine II (ARA II), inhibiteurs calciques, alpha-bloquants, alpha-bêta-bloquants, vasodilatateurs, antihypertenseurs centraux',
+            examens: 'mesure de la pression artérielle, électrocardiogramme, échocardiographie',
+        },
+    ]);
 
-    const handleAddDisease = async (): Promise<void> => {
-        if (!newDisease.trim()) {
+    const [expanded, setExpanded] = useState<number | null>(null);
+    const [isModalVisible, setModalVisible] = useState<boolean>(false);
+    const [isEditModalVisible, setEditModalVisible] = useState<boolean>(false);
+    const [editingIndex, setEditingIndex] = useState<number | null>(null);
+    const [newDisease, setNewDisease] = useState<Disease>({
+        name: '',
+        description: '',
+        symptoms: '',
+        beginDate: '',
+        medications: '',
+        examens: '',
+    });
+    const [editedDisease, setEditedDisease] = useState<Disease>({
+        name: '',
+        description: '',
+        symptoms: '',
+        beginDate: '',
+        medications: '',
+        examens: '',
+    });
+
+    const [selectedYear, setSelectedYear] = useState<number>(2024); 
+    const [selectedMonth, setSelectedMonth] = useState<number>(1); 
+    const [selectedDay, setSelectedDay] = useState<number>(1); 
+    const [editSelectedYear, setEditSelectedYear] = useState<number>(2024); 
+    const [editSelectedMonth, setEditSelectedMonth] = useState<number>(1); 
+    const [editSelectedDay, setEditSelectedDay] = useState<number>(1); 
+
+    // For simple disease addition by profile 
+    const [newDiseaseSimple, setNewDiseaseSimple] = useState<string>('');
+
+    const toggleCard = (index: number): void => { 
+        setExpanded(expanded === index ? null : index);
+    }; 
+
+    // Complex disease management (for the main profile) 
+    const handleAddPress = (): void => {
+        if (
+            !newDisease.name ||
+            !newDisease.description ||
+            !newDisease.symptoms ||
+            !newDisease.medications ||
+            !newDisease.examens
+        ) {
+            Alert.alert('Erreur', 'Veuillez remplir tous les champs pour ajouter une nouvelle maladie.');
+            return;
+        }
+    
+        const newDiseaseData: Disease = {
+            ...newDisease,
+            beginDate: `${selectedDay.toString().padStart(2, '0')}/${selectedMonth.toString().padStart(2, '0')}/${selectedYear}`,
+        };
+    
+        setDiseases([newDiseaseData, ...diseases]);
+        setNewDisease({
+            name: '',
+            description: '',
+            symptoms: '',
+            beginDate: '',
+            medications: '',
+            examens: '',
+        });
+        setModalVisible(false);
+        setSelectedYear(2024);
+        setSelectedMonth(1);
+        setSelectedDay(1); 
+    };
+
+    const handleEditPress = (index: number): void => {
+        const disease = diseases[index];
+        setEditedDisease({ ...disease });
+        setEditingIndex(index);
+        
+        const dateParts = disease.beginDate.split('/');
+        if (dateParts.length === 3) {
+            setEditSelectedDay(parseInt(dateParts[0]));
+            setEditSelectedMonth(parseInt(dateParts[1]));
+            setEditSelectedYear(parseInt(dateParts[2]));
+        }
+        
+        setEditModalVisible(true);
+    };
+
+    const handleSaveEdit = (): void => {
+        if (
+            !editedDisease.name ||
+            !editedDisease.description ||
+            !editedDisease.symptoms ||
+            !editedDisease.medications ||
+            !editedDisease.examens
+        ) {
+            Alert.alert('Erreur', 'Veuillez remplir tous les champs.');
+            return;
+        }
+
+        if (editingIndex !== null) {
+            const updatedDiseases = [...diseases];
+            updatedDiseases[editingIndex] = {
+                ...editedDisease,
+                beginDate: `${editSelectedDay.toString().padStart(2, '0')}/${editSelectedMonth.toString().padStart(2, '0')}/${editSelectedYear}`,
+            };
+            setDiseases(updatedDiseases);
+        }
+
+        setEditModalVisible(false);
+        setEditingIndex(null);
+        Alert.alert('Succès', 'Les informations de la maladie ont été mises à jour.');
+    };
+
+    const handleDeleteDisease = (index: number): void => {
+        const disease = diseases[index];
+        Alert.alert(
+            'Supprimer la maladie',
+            `Êtes-vous sûr de vouloir supprimer "${disease.name}" ?`,
+            [
+                { text: 'Annuler', style: 'cancel' },
+                { 
+                    text: 'Supprimer', 
+                    style: 'destructive',
+                    onPress: () => {
+                        const updatedDiseases = diseases.filter((_, i) => i !== index);
+                        setDiseases(updatedDiseases);
+                    }
+                }
+            ]
+        );
+    };
+
+    // Simple disease management by profile 
+    const handleAddSimpleDisease = async (): Promise<void> => {
+        if (!newDiseaseSimple.trim()) {
             Alert.alert('Erreur', 'Veuillez entrer le nom de la maladie.');
             return;
         }
 
-        const success = await addDisease(newDisease.trim());
+        const success = await addDisease(newDiseaseSimple.trim());
         if (success) {
-            setNewDisease('');
+            setNewDiseaseSimple('');
             setModalVisible(false);
             Alert.alert('Succès', 'Maladie ajoutée avec succès.');
         } else {
@@ -64,7 +220,7 @@ export default function Diseases({ navigation }: DiseasesProps): React.JSX.Eleme
         );
     };
 
-    const getRelationshipText = (relationship?: string) => {
+    const getRelatioinhipText = (relationship?: string) => {
         switch (relationship) {
             case 'self': return 'Mon profil';
             case 'child': return 'Profil enfant';
@@ -73,13 +229,16 @@ export default function Diseases({ navigation }: DiseasesProps): React.JSX.Eleme
             case 'other': return 'Autre profil';
             default: return 'Mon profil';
         }
-    };
+    }; 
 
     React.useLayoutEffect(() => {
         navigation.setOptions({
             title: 'Maladies',
         });
     }, [navigation]);
+
+    // Determine if it's the main profile 
+    const isMainProfile = currentProfile?.name === 'Profil de base' || currentProfile?.relationship === 'self';
 
     return ( 
         <View style={[styles.container, { flex: 1 }]}> 
@@ -93,158 +252,519 @@ export default function Diseases({ navigation }: DiseasesProps): React.JSX.Eleme
                                     {currentProfile.name}
                                 </Text>
                                 <Text style={[styles.cardText, { fontSize: 14, opacity: 0.7 }]}>
-                                    {getRelationshipText(currentProfile.relationship)}
+                                    {getRelatioinhipText(currentProfile.relationship)}
                                 </Text>
                             </View>
                             <TouchableOpacity onPress={() => navigation.navigate('ProfileSelection')}>
                                 <Ionicons name="swap-horizontal" size={24} color={colors.primary} />
                             </TouchableOpacity>
                         </View>
-                    </View> 
-                )} 
-
-                {/* List of diseases */} 
-                {diseases.length > 0 ? (
-                    <View style={[styles.card, { marginBottom: 20}]}> 
-                        <Text style={[styles.cardText, { fontSize: 18, fontWeight: 'bold', marginBottom: 15 }]}> 
-                            Maladies enregistrées ({diseases.length})
-                        </Text>
-                        {diseases.map((disease, index) => ( 
-                            <TouchableOpacity 
-                                key={index} 
-                                style={[styles.card, {
-                                    marginBottom: 10, 
-                                    backgroundColor: colors.background,
-                                    borderWidth: 1,
-                                    borderColor: colors.primary + '30'
-                                }]}
-                            >
-                                <LinearGradient 
-                                    colors={[colors.background, colors.background]} 
-                                    style={[styles.cardGradient, { 
-                                        flexDirection: 'row', 
-                                        justifyContent: 'space-between', 
-                                        alignItems: 'center' 
-                                    }]}
-                                >
-                                    <View style={{ flex: 1 }}>
-                                        <Text style={[styles.cardText, { fontSize: 16, fontWeight: 'bold' }]}>
-                                            {disease}
-                                        </Text>
-                                    </View>
-                                    <TouchableOpacity 
-                                        onPress={() => handleRemoveDisease(disease)}
-                                        style={{ padding: 5 }}
-                                    >
-                                        <Ionicons name="close-circle" size={22} color="#ff6b6b" />
-                                    </TouchableOpacity>
-                                </LinearGradient>
-                            </TouchableOpacity>
-                        ))}
-                    </View>
-                ) : ( 
-                    <View style={[styles.card, { marginBottom: 20, alignItems: 'center', padding: 40 }]}> 
-                        <Ionicons name="medical-outline" size={48} color={colors.iconPrimary} style={{ marginBottom: 15 }} />
-                        <Text style={[styles.cardText, { fontSize: 18, fontWeight: 'bold', textAlign: 'center', marginBottom: 5 }]}> 
-                            Aucune maladie enregistrée
-                        </Text>
-                        <Text style={[styles.cardText, { textAlign: 'center', opacity: 0.7 }]}> 
-                            Ajoutez vos maladies pour un meilleur suivi médical
-                        </Text>
                     </View>
                 )}
 
+                {/* Conditional display based on profile */} 
+                {isMainProfile ? ( 
+                    // For the main profile: predefined complex cards
+                    <>
+                        {diseases.map((disease, index) => (
+                            <TouchableOpacity key={index} onPress={() => toggleCard(index)}>
+                                <View key={index} style={styles.card}>
+                                    <View style={styles.cardHeader}>
+                                        <Text style={styles.cardTitle}>{disease.name}</Text>
+                                        <View style={styles.actionButtons}>
+                                            <TouchableOpacity onPress={() => handleEditPress(index)} style={styles.editButton}>
+                                                <Ionicons name="create-outline" size={25} color={colors.iconPrimary} />
+                                            </TouchableOpacity>
+                                            <TouchableOpacity onPress={() => handleDeleteDisease(index)} style={styles.deleteButton}>
+                                                <Ionicons name="trash-outline" size={25} color="#FF4444" />
+                                            </TouchableOpacity>
+                                        </View>
+                                    </View>
+
+                                    <Text style={styles.cardText}>
+                                        <Text style={styles.bold}>Description: </Text>
+                                        {expanded === index ? disease.description : `${disease.description.slice(0, 70)}...`}
+                                    </Text>
+                                    <Text style={styles.cardText}>
+                                        <Text style={styles.bold}>Symptômes: </Text>
+                                        {expanded === index ? disease.symptoms : `${disease.symptoms.slice(0, 75)}...`}
+                                    </Text>
+                                    <Text style={styles.cardText}>
+                                        <Text style={styles.bold}>Date de début: </Text>
+                                        {disease.beginDate}
+                                    </Text>
+                                    <Text style={styles.cardText}>
+                                        <Text style={styles.bold}>Traitements: </Text>
+                                        {expanded === index ? disease.medications : `${disease.medications.slice(0, 75)}...`}
+                                    </Text>
+                                    <Text style={styles.cardText}>
+                                        <Text style={styles.bold}>Examens: </Text>
+                                        {expanded === index ? disease.examens : `${disease.examens.slice(0, 75)}...`}
+                                    </Text>
+
+                                    <TouchableOpacity onPress={() => toggleCard(index)} style={styles.arrowContainer}>
+                                        <Ionicons
+                                            name={expanded === index ? 'chevron-up-outline' : 'chevron-down-outline'}
+                                            size={24}
+                                            color={colors.iconPrimary}
+                                        />
+                                    </TouchableOpacity>
+                                </View>
+                            </TouchableOpacity>
+                        ))}
+                    </>
+                ) : ( 
+                    // For other profiles: simple list
+                    <>
+                        {profileDiseases && profileDiseases.length > 0 ? (
+                            <View style={[styles.card, { marginBottom: 20}]}> 
+                                <Text style={[styles.cardText, { fontSize: 18, fontWeight: 'bold', marginBottom: 15 }]}> 
+                                    Maladies enregistrées ({profileDiseases.length})
+                                </Text>
+                                {profileDiseases.map((disease, index) => ( 
+                                    <TouchableOpacity key={index} onPress={() => toggleCard(index)}>
+                                        <View key={index} style={styles.card}>
+                                            <View style={styles.cardHeader}>
+                                                <Text style={styles.cardTitle}>{disease}</Text>
+                                            </View>
+                                            <TouchableOpacity onPress={() => handleRemoveDisease(disease)}>
+                                                <Ionicons name="trash-outline" size={22} color="#ff6b6b" />
+                                            </TouchableOpacity>
+                                        </View>
+                                    </TouchableOpacity>
+                                ))}
+                            </View>
+                        ) : (
+                            <View style={[styles.card, { marginBottom: 20, alignItems: 'center', padding: 40 }]}> 
+                                <Ionicons name="medical-outline" size={48} color={colors.iconPrimary} style={{ marginBottom: 15 }} />
+                                <Text style={[styles.cardText, { textAlign: 'center', marginTop: 20 }]}>
+                                    Aucune maladie enregistrée pour ce profil.
+                                </Text>
+                                <Text style={[styles.cardText, { textAlign: 'center', opacity: 0.7 }]}> 
+                                    Ajoutez vos maladies pour un meilleur suivi médical
+                                </Text>
+                            </View>
+                        )}
+                    </> 
+                )}
+
                 {/* Button to add a new disease */} 
-                <TouchableOpacity 
-                    onPress={() => setModalVisible(true)} 
-                    style={[styles.card]}
-                >
-                    <LinearGradient 
-                        colors={[colors.primary, colors.secondary]}
-                        style={[styles.cardGradient, { alignItems: 'center', padding: 20 }]} 
-                    >
-                        <View style={{ flexDirection: 'row', alignItems: 'center'}}>
-                            <Ionicons name="add-circle-outline" size={24} color="#fff" style={{ marginRight: 10 }} />
-                            <Text style={[styles.cardText, { color: '#fff', fontWeight: 'bold', fontSize: 16 }]}>
-                                Ajouter une maladie
-                            </Text>
-                        </View>
-                    </LinearGradient>
-                </TouchableOpacity>
+                <View style={styles.buttonContainer}>
+                    <TouchableOpacity style={styles.button} onPress={() => setModalVisible(true)}>
+                        <LinearGradient colors={[colors.primary, colors.secondary]} style={styles.gradient}>
+                            <Text style={styles.buttonText}>Ajouter</Text>
+                        </LinearGradient>
+                    </TouchableOpacity>
+                    <TouchableOpacity style={styles.button} onPress={() => navigation.goBack()}>
+                        <LinearGradient colors={[colors.primary, colors.secondary]} style={styles.gradient}>
+                            <Text style={styles.buttonText}>Retour</Text>
+                        </LinearGradient>
+                    </TouchableOpacity>
+                </View>
             </ScrollView>
 
-            {/* Modal to add a disease */}
-            <Modal visible={isModalVisible} animationType="slide" transparent={true}>
-                <View style={{
-                    flex: 1,
-                    backgroundColor: 'rgba(0,0,0,0.5)',
-                    justifyContent: 'center',
-                    alignItems: 'center',
-                    padding: 20
-                }}>
-                    <View style={{
-                        backgroundColor: colors.background,
-                        borderRadius: 15,
-                        padding: 20,
-                        width: '100%',
-                        maxWidth: 400
-                    }}>
-                        <Text style={[styles.cardText, { fontSize: 18, fontWeight: 'bold', textAlign: 'center', marginBottom: 20 }]}>
-                            Ajouter une maladie
+            {/* Conditional modal for adding a new disease */} 
+            <Modal visible={isModalVisible} animationType="slide">
+                {isMainProfile ? (
+                    // Complex modal for the main profile
+                    <ScrollView
+                        contentContainerStyle={{
+                            flexGrow: 1,
+                            backgroundColor: colors.background,
+                            padding: 20
+                        }}
+                        keyboardShouldPersistTaps="handled"
+                    > 
+                        <Text style={[styles.cardText, { fontSize: 20, fontWeight: 'bold', textAlign: 'center', marginBottom: 20 }]}> 
+                            Ajouter une maladie 
                         </Text>
 
-                        <Text style={[styles.cardText, { marginBottom: 10 }]}>Nom de la maladie</Text>
-                        <TextInput
+                        <TextInput 
+                            placeholder="Nom" 
+                            value={newDisease.name}
+                            onChangeText={(text) => setNewDisease({ ...newDisease, name: text })}
                             style={{
                                 borderWidth: 1,
                                 borderColor: colors.primary,
                                 borderRadius: 10,
                                 padding: 15,
-                                marginBottom: 20,
+                                marginBottom: 15,
                                 fontSize: 16,
                                 color: colors.text
                             }}
-                            value={newDisease}
-                            onChangeText={setNewDisease}
-                            placeholder="Ex: Diabète, Hypertension, Asthme..."
-                            placeholderTextColor={colors.text + '80'}
+                            placeholderTextColor={colors.text + '80'} 
                         />
 
-                        <View style={{ flexDirection: 'row', justifyContent: 'space-between' }}>
-                            <TouchableOpacity
-                                style={{
-                                    flex: 1,
-                                    padding: 15,
-                                    borderRadius: 10,
+                        <TextInput 
+                            placeholder="Description" 
+                            value={newDisease.description}
+                            onChangeText={(text) => setNewDisease({ ...newDisease, description: text })}
+                            style={{
+                                borderWidth: 1,
+                                borderColor: colors.primary,
+                                borderRadius: 10,
+                                padding: 15,
+                                marginBottom: 15,
+                                fontSize: 16,
+                                color: colors.text,
+                                minHeight: 80
+                            }}
+                            multiline
+                            placeholderTextColor={colors.text + '80'} 
+                        />
+                        
+                        <TextInput 
+                            placeholder="Symptômes"
+                            value={newDisease.symptoms}
+                            onChangeText={(text) => setNewDisease({ ...newDisease, symptoms: text })}
+                            style={{
+                                borderWidth: 1,
+                                borderColor: colors.primary,
+                                borderRadius: 10,
+                                padding: 15,
+                                marginBottom: 15,
+                                fontSize: 16,
+                                color: colors.text,
+                                minHeight: 80
+                            }}
+                            multiline
+                            placeholderTextColor={colors.text + '80'} 
+                        />
+
+                        <View style={{ flexDirection: 'row', justifyContent: 'space-between', marginBottom: 15 }}> 
+                            <View style={{ flex: 1, marginRight: 5 }}> 
+                                <CustomPicker 
+                                    label="Jour" 
+                                    selectedValue={selectedDay} 
+                                    onValueChange={(value) => setSelectedDay(Number(value))}
+                                    options={Array.from({ length: 31 }, (_, i) => ({ label: (i + 1).toString(), value: (i + 1).toString() }))}
+                                    placeholder="01"
+                                />
+                            </View>
+                            <View style={{ flex: 1, marginHorizontal: 5 }}> 
+                                <CustomPicker 
+                                    label="Mois" 
+                                    selectedValue={selectedMonth} 
+                                    onValueChange={(value) => setSelectedMonth(Number(value))}
+                                    options={Array.from({ length: 12 }, (_, i) => ({ label: (i + 1).toString(), value: (i + 1).toString() }))}
+                                    placeholder="01"
+                                />
+                            </View>
+                            <View style={{ flex: 1, marginLeft: 5 }}> 
+                                <CustomPicker 
+                                    label="Année" 
+                                    selectedValue={selectedYear} 
+                                    onValueChange={(value) => setSelectedYear(Number(value))}
+                                    options={Array.from({ length: 100 }, (_, i) => ({ label: (i + 1920).toString(), value: (i + 1920).toString() }))}
+                                    placeholder="2024"
+                                />
+                            </View>
+                        </View>
+
+                        <TextInput 
+                            placeholder="Traitements"
+                            value={newDisease.medications}
+                            onChangeText={(text) => setNewDisease({ ...newDisease, medications: text })}
+                            style={{
+                                borderWidth: 1,
+                                borderColor: colors.primary,
+                                borderRadius: 10,
+                                padding: 15,
+                                marginBottom: 15,
+                                fontSize: 16,
+                                color: colors.text,
+                                minHeight: 80
+                            }}
+                            multiline
+                            placeholderTextColor={colors.text + '80'} 
+                        />
+
+                        <TextInput 
+                            placeholder="Examens" 
+                            value={newDisease.examens}
+                            onChangeText={(text) => setNewDisease({ ...newDisease, examens: text })}
+                            style={{
+                                borderWidth: 1,
+                                borderColor: colors.primary,
+                                borderRadius: 10,
+                                padding: 15,
+                                marginBottom: 30,
+                                fontSize: 16,
+                                color: colors.text,
+                                minHeight: 80
+                            }}
+                            multiline
+                            placeholderTextColor={colors.text + '80'} 
+                        />
+
+                        <View style={{ flexDirection: 'row', justifyContent: 'space-between' }}> 
+                            <TouchableOpacity 
+                                style={{ 
+                                    flex: 1, 
+                                    padding: 15, 
+                                    borderRadius: 10, 
                                     backgroundColor: colors.secondary + '30',
                                     marginRight: 10,
                                     alignItems: 'center'
                                 }}
-                                onPress={() => {
-                                    setModalVisible(false);
-                                    setNewDisease('');
-                                }}
+                                onPress={() => setModalVisible(false)}
                             >
                                 <Text style={{ color: colors.text, fontWeight: 'bold' }}>Annuler</Text>
                             </TouchableOpacity>
-
-                            <TouchableOpacity
-                                style={{
-                                    flex: 1,
-                                    padding: 15,
-                                    borderRadius: 10,
+                            <TouchableOpacity 
+                                style={{ 
+                                    flex: 1, 
+                                    padding: 15, 
+                                    borderRadius: 10, 
                                     backgroundColor: colors.primary,
-                                    marginLeft: 10,
+                                    marginLeft: 10, 
                                     alignItems: 'center'
                                 }}
-                                onPress={handleAddDisease}
+                                onPress={handleAddPress}
                             >
                                 <Text style={{ color: '#fff', fontWeight: 'bold' }}>Ajouter</Text>
                             </TouchableOpacity>
                         </View>
+                    </ScrollView>
+                ) : ( 
+                    // Simple modal for other profiles 
+                    <View style={{ 
+                        flex: 1, 
+                        backgroundColor: 'rgba(0, 0, 0, 0.5)',
+                        justifyContent: 'center',
+                        alignItems: 'center',
+                        padding: 20
+                    }}> 
+                        <View style={{ 
+                            backgroundColor: colors.background,
+                            borderRadius: 15,
+                            padding: 20,
+                            width: '100%',
+                            maxWidth: 400
+                        }}> 
+                            <Text style={[styles.cardText, { fontSize: 18, fontWeight: 'bold', textAlign: 'center', marginBottom: 20 }]}> 
+                                Ajouter une maladie 
+                            </Text>
+
+                            <Text style={[styles.cardText, { marginBottom: 10 }]}> Nom de la maladie</Text>
+                            <TextInput 
+                                style={{ 
+                                    borderWidth: 1,
+                                    borderColor: colors.primary,
+                                    borderRadius: 10,
+                                    padding: 15,
+                                    marginBottom: 20,
+                                    fontSize: 16,
+                                    color: colors.text
+                                }}
+                                value={newDiseaseSimple}
+                                onChangeText={setNewDiseaseSimple}
+                                placeholder="Ex: Diabète, Hypertension, Asthme..."
+                                placeholderTextColor={colors.text + '80'}
+                            />
+
+                            <View style={{ flexDirection: 'row', justifyContent: 'space-between' }}> 
+                                <TouchableOpacity 
+                                    style={{ 
+                                        flex: 1, 
+                                        padding: 15, 
+                                        borderRadius: 10, 
+                                        backgroundColor: colors.secondary + '30',
+                                        marginRight: 10,
+                                        alignItems: 'center'
+                                    }}
+                                    onPress={() => { 
+                                        setModalVisible(false); 
+                                        setNewDiseaseSimple('');
+                                    }}
+                                > 
+                                    <Text style={{ color: colors.text, fontWeight: 'bold' }}> Annuler</Text> 
+                                </TouchableOpacity>
+
+                                <TouchableOpacity 
+                                    style={{ 
+                                        flex: 1, 
+                                        padding: 15, 
+                                        borderRadius: 10, 
+                                        backgroundColor: colors.primary,
+                                        marginLeft: 10,
+                                        alignItems: 'center'
+                                    }}
+                                    onPress={handleAddSimpleDisease}
+                                >
+                                    <Text style={{ color: '#fff', fontWeight: 'bold' }}> Ajouter</Text> 
+                                </TouchableOpacity>
+                            </View>
+                        </View>
                     </View>
-                </View>
+                )}
             </Modal>
-        </View>
+
+            {/* Editing modal for the main profile */} 
+            {isMainProfile && ( 
+                <Modal visible={isEditModalVisible} animationType="slide"> 
+                    <ScrollView 
+                        contentContainerStyle={{ 
+                            flexGrow: 1, 
+                            backgroundColor: colors.background,
+                            padding: 20
+                        }}
+                        keyboardShouldPersistTaps="handled"
+                    >
+                        <Text style={[styles.cardText, { fontSize: 20, fontWeight: 'bold', textAlign: 'center', marginBottom: 20 }]}> 
+                            Modifier la maladie 
+                        </Text>
+
+                        <TextInput 
+                            placeholder="Nom" 
+                            value={editedDisease.name}
+                            onChangeText={(text) => setEditedDisease({ ...editedDisease, name: text })}
+                            style={{
+                                borderWidth: 1,
+                                borderColor: colors.primary,
+                                borderRadius: 10,
+                                padding: 15,
+                                marginBottom: 15,
+                                fontSize: 16,
+                                color: colors.text
+                            }}
+                            placeholderTextColor={colors.text + '80'}
+                        />
+
+                        <TextInput 
+                            placeholder="Description" 
+                            value={editedDisease.description}
+                            onChangeText={(text) => setEditedDisease({ ...editedDisease, description: text })}
+                            style={{
+                                borderWidth: 1,
+                                borderColor: colors.primary,
+                                borderRadius: 10,
+                                padding: 15,
+                                marginBottom: 15,
+                                fontSize: 16,
+                                color: colors.text,
+                                minHeight: 80
+                            }}
+                            multiline
+                            placeholderTextColor={colors.text + '80'}
+                        />
+
+                        <TextInput 
+                            placeholder="Symptômes" 
+                            value={editedDisease.symptoms}
+                            onChangeText={(text) => setEditedDisease({ ...editedDisease, symptoms: text })}
+                            style={{
+                                borderWidth: 1,
+                                borderColor: colors.primary,
+                                borderRadius: 10,
+                                padding: 15,
+                                marginBottom: 15,
+                                fontSize: 16,
+                                color: colors.text,
+                                minHeight: 80
+                            }}
+                            multiline
+                            placeholderTextColor={colors.text + '80'}
+                        />
+
+                        <View style={{ flexDirection: 'row', justifyContent: 'space-between', marginBottom: 15 }}> 
+                            <View style={{ flex: 1, marginRight: 5 }}> 
+                                <CustomPicker 
+                                    label="Jour" 
+                                    selectedValue={editSelectedDay} 
+                                    onValueChange={(value) => setEditSelectedDay(Number(value))}
+                                    options={Array.from({ length: 31 }, (_, i) => ({ label: (i + 1).toString(), value: i + 1}))}
+                                    placeholder="01"
+                                />
+                            </View>
+                            <View style={{ flex: 1, marginHorizontal: 5 }}> 
+                                <CustomPicker 
+                                    label="Mois" 
+                                    selectedValue={editSelectedMonth} 
+                                    onValueChange={(value) => setEditSelectedMonth(Number(value))}
+                                    options={Array.from({ length: 12 }, (_, i) => ({ label: (i + 1).toString(), value: i + 1 }))}
+                                    placeholder="01"
+                                />
+                            </View>
+                            <View style={{ flex: 1, marginLeft: 5 }}> 
+                                <CustomPicker 
+                                    label="Année" 
+                                    selectedValue={editSelectedYear} 
+                                    onValueChange={(value) => setEditSelectedYear(Number(value))}
+                                    options={Array.from({ length: 100 }, (_, i) => ({ label: (1980 + i).toString(), value: 1980 + i }))}
+                                    placeholder="2024"
+                                />
+                            </View>
+                        </View>
+
+                        <TextInput 
+                            placeholder="Traitements" 
+                            value={editedDisease.medications}
+                            onChangeText={(text) => setEditedDisease({ ...editedDisease, medications: text })}
+                            style={{
+                                borderWidth: 1,
+                                borderColor: colors.primary,
+                                borderRadius: 10,
+                                padding: 15,
+                                marginBottom: 15,
+                                fontSize: 16,
+                                color: colors.text,
+                                minHeight: 80
+                            }}
+                            multiline
+                            placeholderTextColor={colors.text + '80'}
+                        />
+
+                        <TextInput 
+                            placeholder="Examens" 
+                            value={editedDisease.examens}
+                            onChangeText={(text) => setEditedDisease({ ...editedDisease, examens: text })}
+                            style={{
+                                borderWidth: 1,
+                                borderColor: colors.primary,
+                                borderRadius: 10,
+                                padding: 15,
+                                marginBottom: 30,
+                                fontSize: 16,
+                                color: colors.text,
+                                minHeight: 80
+                            }}
+                            multiline
+                            placeholderTextColor={colors.text + '80'}
+                        />
+
+                        <View style={{ flexDirection: 'row', justifyContent: 'space-between' }}> 
+                            <TouchableOpacity 
+                                style={{ 
+                                    flex: 1, 
+                                    padding: 15, 
+                                    borderRadius: 10, 
+                                    backgroundColor: colors.secondary + '30',
+                                    marginRight: 10,
+                                    alignItems: 'center'
+                                }}
+                                onPress={() => setEditModalVisible(false)}
+                            >
+                                <Text style={{ color: colors.text, fontWeight: 'bold' }}>Annuler</Text>
+                            </TouchableOpacity>
+                            <TouchableOpacity 
+                                style={{ 
+                                    flex: 1, 
+                                    padding: 15, 
+                                    borderRadius: 10, 
+                                    backgroundColor: colors.primary,
+                                    marginLeft: 10,
+                                    alignItems: 'center'
+                                }}
+                                onPress={handleSaveEdit}
+                            >
+                                <Text style={{ color: '#fff', fontWeight: 'bold' }}>Enregistrer</Text>
+                            </TouchableOpacity>
+                        </View>
+
+                    </ScrollView>
+                </Modal>
+            )}
+        </View> 
     ); 
 }
